// src/interfaces/appInterfaces.ts
export interface LoginData {
  correo: string;
  password: string;
}

export interface LoginResponse {
  usuario: User;
  token: string;
}

export interface User {
  rol: string;
  estado: boolean;
  google: boolean;
  nombre: string;
  correo: string;
  uid: string;
  img?: string;
}

export enum UserStatus {
  CHECKING,
  AUTHENTICATED,
  NOT_AUTHENTICATED,
}